+++ js/main.js
@@ -1,2 +1,3 @@
-import './create-photos-description.js';
+import './CREATE_PHOTOS_DESCRIPTION.js';
 import './line-size-check.js';
+import './miniatures-rendering.js';
